% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\docType{data}
\name{bloodpressuremeasurements}
\alias{bloodpressuremeasurements}
\title{Sample Blood Pressure measurement data}
\format{
\tabular{ll}{
\bold{name} \tab \bold{description} \cr
country    \tab Date    \cr
  year       \tab Time    \cr
  population \tab Category\cr
  country    \tab Sys     \cr
  year       \tab Dia     \cr
  population \tab Pulse   
}
}
\usage{
bloodpressuremeasurements
}
\description{
A subset of data to test the package
}
\keyword{datasets}
