% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graph.R
\name{filter_and_draw}
\alias{filter_and_draw}
\title{Filter and plot blood pressure readings}
\usage{
filter_and_draw(start, end)
}
\arguments{
\item{start}{A date.}

\item{end}{A date.}
}
\value{
plot.
}
\description{
Filter and plot blood pressure readings
}
